+++ R/CytoPipelineClass.R
@@ -51,9 +51,9 @@
 #' ### *** EXAMPLE 1: building CytoPipeline step by step *** ###
 #' 
 #' rawDataDir <-
-#'     paste0(system.file("extdata", package = "CytoPipeline"), "/")
+#'     system.file("extdata", package = "CytoPipeline")
 #' experimentName <- "OMIP021_PeacoQC"
-#' sampleFiles <- paste0(rawDataDir, list.files(rawDataDir,
+#' sampleFiles <- file.path(rawDataDir, list.files(rawDataDir,
 #'                                              pattern = "sample_"))
 #'                                              
 #' outputDir <- base::tempdir()
@@ -223,7 +223,7 @@
 #' ### *** EXAMPLE 2: building CytoPipeline from JSON file *** ###
 #' 
 #' jsonDir <- system.file("extdata", package = "CytoPipeline")
-#' jsonPath <- paste0(jsonDir, "/pipelineParams.json")
+#' jsonPath <- file.path(jsonDir, "pipelineParams.json")
 #' 
 #' # Note that the `experimentName` and `sampleFiles` are here 
 #' # specified in the JSON file itself. This is not necessary, as