+++ tests/testthat/test-utils.R
@@ -419,7 +419,7 @@ test_that("writeFlowFrame works", {
     
     ff_c <- readRDS(test_path("fixtures", "ff_c.rds"))
     
-    expect_error(writeFlowFrame(ff, dir = paste0(outputDir, "/notThere")),
+    expect_error(writeFlowFrame(ff, dir = file.path(outputDir, "notThere")),
                  regexp = "Provided directory does not exist")
     
     prefix <- "File_"
@@ -431,11 +431,8 @@ test_that("writeFlowFrame works", {
                    suffix = suffix, 
                    format = "fcs")
     
-    outputFile <- paste0(outputDir, "/", 
-                         prefix, 
-                         "Donor1",
-                         suffix,
-                         ".fcs")
+    outputFile <- file.path(outputDir, 
+                            paste0(prefix, "Donor1", suffix, ".fcs"))
     
     thisFF <- flowCore::read.FCS(outputFile, transform = FALSE)
     expect_true(all(round(flowCore::exprs(thisFF), 0)
@@ -447,10 +444,8 @@ test_that("writeFlowFrame works", {
                    suffix = suffix, 
                    format = "csv")
     
-    outputCSV <- paste0(outputDir, "/", 
-                        prefix,
-                        suffix,
-                        ".csv")
+    outputCSV <- file.path(outputDir, 
+                           paste0(prefix, suffix, ".csv"))
     
     thisExpr <- read.csv(file = outputCSV)
     expect_true(all(round(thisExpr,4) == round(flowCore::exprs(ff_c), 4)))